chore: remove wandb_helper.py file
